requirements.txt
pydantic-ai
openai
qdrant-client
fastapi
python-multipart
pypdf



google-generativeai


